use the actual vector size in control value conversion
Dynamic Spans have a maximum extent ('dynamic_extent') but their associated
'ControlInfo' can contain a variable number of elements. For dynamic Spans
with a single element in the default 'ControlValue', this previously caused
the exception 'std::length_error' since a vector with "maximum extent"
elements was constructed.

Fix this by ignoring the extent and using the actual number of elements.