check for supported parameter value type based on control id and not value
libcamera 'Control<T>' and their related 'ControlInfo' do not necessarily
have the same control and value types. This previously caused issues when
an unsupported control type, such as a span of a complex type, is mapped
to a ROS parameter via the type of the default 'ControlValue'.

Fix this by using the the actual control type, regardless of the types of
the values in the 'ControlInfo'.