support scalar min/max bounds for Span controls
In libcamera 0.6, some controls changed their default value from a scalar
to an array. E.g. 'FrameDurationLimits' for the 'imx708_wide' changed from
the scalar '33333' to an array '[ 33333, 33333 ]'. That breaks the array
value clamping in 'clamp_array', as this assumes that default, minimum and
maximum value are of the same type and dimensionality.

Fix this by clamping to scalar min/max bounds for arrays, when their bounds
are scalars.